+++ Tests/Set-CIisAppPoolPeriodicRestart.Tests.ps1
@@ -101,7 +101,7 @@ BeforeAll {
     }
 }
 
-Describe 'Set-CIisAppPoolRecyclingPeriodicRestart' {
+Describe 'Set-CIisAppPoolPeriodicRestart' {
     BeforeAll {
         Start-W3ServiceTestFixture
     }
@@ -111,66 +111,66 @@ Describe 'Set-CIisAppPoolRecyclingPeriodicRestart' {
     }
 
     BeforeEach {
-        $script:appPoolName = "Set-CIisAppPoolRecyclingPeriodicRestart$($script:testNum++)"
-        Set-CIisAppPoolRecyclingPeriodicRestart -AsDefaults @script:defaultDefaults
+        $script:appPoolName = "Set-CIisAppPoolPeriodicRestart$($script:testNum++)"
+        Set-CIisAppPoolPeriodicRestart -AsDefaults @script:defaultDefaults
         Install-CIisAppPool -Name $script:appPoolName
     }
 
     AfterEach {
         Uninstall-CIisAppPool -Name $script:appPoolName
-        Set-CIisAppPoolRecyclingPeriodicRestart -AsDefaults @script:defaultDefaults
+        Set-CIisAppPoolPeriodicRestart -AsDefaults @script:defaultDefaults
     }
 
     It 'should set and reset all values' {
         $infos = @()
-        Set-CIisAppPoolRecyclingPeriodicRestart -AppPoolName $script:appPoolName `
-                                                @script:nonDefaultArgs `
-                                                -Schedule '01:00:00', '13:00:00' `
-                                                -InformationVariable 'infos'
+        Set-CIisAppPoolPeriodicRestart -AppPoolName $script:appPoolName `
+                                       @script:nonDefaultArgs `
+                                       -Schedule '01:00:00', '13:00:00' `
+                                       -InformationVariable 'infos'
         $infos | Should -Not -BeNullOrEmpty
         ThenHasValues $script:nonDefaultArgs -AndSchedule '01:00:00', '13:00:00'
 
         # Make sure no information messages get written because no changes are being made.
-        Set-CIisAppPoolRecyclingPeriodicRestart -AppPoolName $script:appPoolName `
-                                                @script:nonDefaultArgs `
-                                                -Schedule '01:00:00', '13:00:00' `
-                                                -InformationVariable 'infos'
+        Set-CIisAppPoolPeriodicRestart -AppPoolName $script:appPoolName `
+                                       @script:nonDefaultArgs `
+                                       -Schedule '01:00:00', '13:00:00' `
+                                       -InformationVariable 'infos'
         $infos | Should -BeNullOrEmpty
         ThenHasValues $script:nonDefaultArgs -AndSchedule '01:00:00', '13:00:00'
 
-        Set-CIisAppPoolRecyclingPeriodicRestart -AppPoolName $script:appPoolName
+        Set-CIisAppPoolPeriodicRestart -AppPoolName $script:appPoolName
         ThenHasDefaultValues
     }
 
     It 'should support WhatIf when updating all values' {
-        Set-CIisAppPoolRecyclingPeriodicRestart -AppPoolName $script:appPoolName `
-                                                @script:nonDefaultArgs `
-                                                -Schedule '12:34:00', '23:45:00' `
-                                                -WhatIf
+        Set-CIisAppPoolPeriodicRestart -AppPoolName $script:appPoolName `
+                                       @script:nonDefaultArgs `
+                                       -Schedule '12:34:00', '23:45:00' `
+                                       -WhatIf
         ThenHasDefaultValues
     }
 
     It 'should support WhatIf when resetting all values back to defaults' {
-        Set-CIisAppPoolRecyclingPeriodicRestart -AppPoolName $script:appPoolName @script:nonDefaultArgs
+        Set-CIisAppPoolPeriodicRestart -AppPoolName $script:appPoolName @script:nonDefaultArgs
         ThenHasValues $script:nonDefaultArgs
-        Set-CIisAppPoolRecyclingPeriodicRestart -AppPoolName $script:appPoolName -WhatIf
+        Set-CIisAppPoolPeriodicRestart -AppPoolName $script:appPoolName -WhatIf
         ThenHasValues $script:nonDefaultArgs
     }
 
     It 'should change values and reset to defaults' {
-        Set-CIisAppPoolRecyclingPeriodicRestart -AppPoolName $script:appPoolName @script:nonDefaultArgs -ErrorAction Ignore
+        Set-CIisAppPoolPeriodicRestart -AppPoolName $script:appPoolName @script:nonDefaultArgs -ErrorAction Ignore
         ThenHasValues $script:nonDefaultArgs
 
         $someArgs = @{
             'memory' = 6000000;
             'privateMemory' = 7000000;
         }
-        Set-CIisAppPoolRecyclingPeriodicRestart -AppPoolName $script:appPoolName @someArgs
+        Set-CIisAppPoolPeriodicRestart -AppPoolName $script:appPoolName @someArgs
         ThenHasValues $someArgs
     }
 
     It 'should change default settings' {
-        Set-CIisAppPoolRecyclingPeriodicRestart -AsDefaults @script:nonDefaultArgs
+        Set-CIisAppPoolPeriodicRestart -AsDefaults @script:nonDefaultArgs
         ThenDefaultsSetTo @script:nonDefaultArgs
     }
 }