+++ CHANGELOG.md
@@ -107,7 +107,7 @@ CPU settings.
 the objects returned by the Carbon.IIS module.
 * Function `Set-CIisWebsiteLogFile` for configuring a website's log file settings and the website defaults log file
 settings.
-* Function `Set-CIisAppPoolRecyclingPeriodicRestart` for configuring an application pool's periodic restart settings and
+* Function `Set-CIisAppPoolPeriodicRestart` for configuring an application pool's periodic restart settings and
 the application pool defaults periodic restart settings.
 * `Get-CIIsAppPool` and `Get-CIisWebsite` can now return the application pool defaults and the website
 defaults settings, respectively. Use the new `AsDefaults` switch.